chapter07 : taylor sine approx. wip !

+++ pap_odin/part_four/chapter07_approximation_using_taylor_series.odin
@@ -6,9 +6,9 @@ import "core:math"
 import "../pap_common"
 
 @(private="file")
-factorial :: proc(x: u32) -> u32 {
-	input: u32 = x
-	result: u32 = input
+factorial :: proc(x: u64) -> u64 {
+	input: u64 = x
+	result: u64 = input
     for input > 1
     {
         input -= 1
@@ -19,7 +19,7 @@ factorial :: proc(x: u32) -> u32 {
 }
 
 @(private="file")
-taylor_sine_coefficient :: proc(power: u32) -> f64 {
+taylor_sine_coefficient :: proc(power: u64) -> f64 {
     sign: f64 = ((((power - 1) / 2) % 2) == 1) ? -1.0 : 1.0
     result: f64 = (sign / cast(f64)factorial(power))
 
@@ -27,12 +27,12 @@ taylor_sine_coefficient :: proc(power: u32) -> f64 {
 }
 
 @(private="file")
-taylor_sine_approximation :: proc(max_power: u32, x: f64) -> f64 {
+taylor_sine_approximation :: proc(max_power: u64, x: f64) -> f64 {
     result: f64 = 0
     
     x_squared: f64 = x * x
     x_power: f64 = x
-    for power: u32 = 1; power <= max_power; power += 2 {
+    for power: u64 = 1; power <= max_power; power += 2 {
         result += x_power * taylor_sine_coefficient(power)
         x_power *= x_squared
     }
@@ -46,7 +46,7 @@ chapter07_approximation_using_taylor_series :: proc() {
     tester: Math_Op_Tester
 
 	fmt.printfln("Range : [%+.24f, %+.24f]", 0.0, PI64/2)
-    for power: u32 = 1; power <= 31; power += 2 {
+    for power: u64 = 1; power <= 31; power += 2 {
 		for math_op_precision_tester_try_setup_next_precision_test(&tester, 0.0, PI64/2) {
 			reference_value: f64 = math.sin_f64(tester.input_value)
 	        math_op_precision_tester_test_using_latest_precision_test(&tester,
